+++ packages/reakit-utils/CHANGELOG.md
@@ -3,6 +3,25 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [0.7.0](https://github.com/reakit/reakit/tree/master/packages/reakit-utils/compare/reakit-utils@0.6.7...reakit-utils@0.7.0) (2019-11-08)
+
+**Note:** Version bump only for package reakit-utils
+
+
+
+
+
+## [0.6.7](https://github.com/reakit/reakit/tree/master/packages/reakit-utils/compare/reakit-utils@0.6.6...reakit-utils@0.6.7) (2019-11-02)
+
+
+### Bug Fixes
+
+* Fix `Tabbable` focus behavior on Mac Safari/Firefox ([#458](https://github.com/reakit/reakit/tree/master/packages/reakit-utils/issues/458)) ([8306241](https://github.com/reakit/reakit/tree/master/packages/reakit-utils/commit/8306241))
+
+
+
+
+
 ## [0.6.6](https://github.com/reakit/reakit/tree/master/packages/reakit-utils/compare/reakit-utils@0.6.5...reakit-utils@0.6.6) (2019-10-12)